organized print methods into own file and organized all files in src and include directories
