--- Trying_to_build/dfa_true_output.dot
@@ -1,12 +1,0 @@
-digraph DFA {
-    node [shape = point] START;
-    node [shape = circle] ;
-    node [shape = doublecircle] q0 q1 q2 ;
-    START -> q0 [label="START"];
-    q0 -> q0 [label="0"];
-    q0 -> q1 [label="1"];
-    q1 -> q2 [label="0"];
-    q1 -> q1 [label="1"];
-    q2 -> q2 [label="0"];
-    q2 -> q2 [label="1"];
-}
