--- dfa.dot
@@ -1,18 +1,0 @@
-digraph DFA {
-node [shape = point] START;
-node [shape = circle] q0 q1 q2 ;
-node [shape = doublecircle] q3 q4 q5 ;
-START -> q0 [label="START"];
-q0 -> q1 [label="b"];
-q0 -> q2 [label="a"];
-q1 -> q1 [label="b"];
-q1 -> q1 [label="a"];
-q2 -> q3 [label="b"];
-q2 -> q1 [label="a"];
-q3 -> q4 [label="b"];
-q3 -> q5 [label="a"];
-q4 -> q4 [label="b"];
-q4 -> q5 [label="a"];
-q5 -> q4 [label="b"];
-q5 -> q5 [label="a"];
-}
