+++ include/nfa.h
