--- nfa.dot
@@ -1,20 +1,0 @@
-digraph nfa{
-node[shape = point] START
-node[shape = circle] q0 q1 q2 q3 q4 q5 q6 q7 q8 q9 q10 
-node[shape = doublecircle] q11
-START -> q0[label = START]
-q0 -> q1[label = "a"]
-q1 -> q2[label = "$"]
-q2 -> q3[label = "b"]
-q3 -> q10[label = "$"]
-q10 -> q8[label = "$"]
-q10 -> q11[label = "$"]
-q8 -> q4[label = "$"]
-q8 -> q6[label = "$"]
-q6 -> q7[label = "b"]
-q7 -> q9[label = "$"]
-q9 -> q8[label = "$"]
-q9 -> q11[label = "$"]
-q4 -> q5[label = "a"]
-q5 -> q9[label = "$"]
-}
