+++ src/dfa.cpp
@@ -1,6 +1,5 @@
 #include "dfa.h"
 
-#include <fstream>
 #include <iomanip>
 #include <iostream>
 #include <queue>
@@ -8,7 +7,7 @@
 #include <unordered_set>
 #include <vector>
 
-#include "nfa.h"
+#include "print_utils.h"
 
 // Constructor definition
 DFA::DFA(
@@ -20,6 +19,8 @@ DFA::DFA(
       transitionFn(transitionFn),
       startState(startState),
       finalStates(finalStates) {}
+
+// Function to find epsilon closure
 
 std::unordered_map<int, std::unordered_set<int>> DFA::findeClosure(NFA &nfa) {
   std::unordered_map<int, std::unordered_set<int>> eClosure;
@@ -46,30 +47,7 @@ std::unordered_map<int, std::unordered_set<int>> DFA::findeClosure(NFA &nfa) {
   return eClosure;
 }
 
-void DFA::printeClosure(
-    NFA &nfa, std::unordered_map<int, std::unordered_set<int>> &eClosure) {
-  std::cout << "Epsilon Closures are:";
-  for (int state : nfa.states) {
-    std::cout << "\nState " << state << ": ";
-    for (int c : eClosure[state]) {
-      std::cout << c << " ";
-    }
-  }
-}
-
-void DFA::printMappings(
-    std::unordered_map<int, std::unordered_set<int>> &mappedStates) {
-  std::cout << "\n\nMapped states are:\n";
-  std::cout << "DFA == NFA\n";
-  int n = mappedStates.size();
-  for (int i = 0; i < n; i++) {
-    std::cout << i << " == ";
-    for (auto c : mappedStates[i]) {
-      std::cout << c << ", ";
-    }
-    std::cout << "\b\b \n";
-  }
-}
+// Function to convert NFA to DFA
 
 DFA DFA::convertNFAtoDFA(NFA &nfa) {
   std::unordered_map<int, std::unordered_set<int>> eClosure = findeClosure(nfa);
@@ -133,85 +111,4 @@ DFA DFA::convertNFAtoDFA(NFA &nfa) {
   }
 
   return dfa;
-}
-
-void DFA::printNFA(NFA &nfa) {
-  int width = 6;
-  std::cout << "Given NFA:\n";
-  std::cout << "States: ";
-  for (int c : nfa.states) {
-    std::cout << c << ", ";
-  }
-
-  std::cout << "\b\b \nInput symbols: ";
-  for (char c : nfa.inputSymbols) {
-    std::cout << c << ", ";
-  }
-
-  std::cout << "\b\b \nTransition Function:\n";
-  std::cout << std::setw(width) << "";
-  for (char c : nfa.inputSymbols) {
-    std::cout << std::setw(width) << c;
-  }
-  std::cout << std::setw(width) << "$" << std::endl;
-  for (int srcState : nfa.states) {
-    std::cout << std::setw(width) << srcState;
-    std::vector<char> symbols = nfa.inputSymbols;
-    symbols.push_back('$');
-    for (char symbol : symbols) {
-      std::vector<int> destStates = nfa.transitionFn[srcState][symbol];
-      std::string str = "";
-      for (int state : destStates) {
-        str += std::to_string(state);
-      }
-      std::cout << std::setw(width) << str;
-    }
-    std::vector<int> destStates = nfa.transitionFn[srcState]['$'];
-    std::string str = "";
-    for (int state : destStates) {
-      str += state;
-    }
-    std::cout << std::setw(width) << str << std::endl;
-  }
-
-  std::cout << "Start State: " << nfa.startState;
-  std::cout << "\nFinal States: ";
-  for (int c : nfa.finalStates) {
-    std::cout << c << ", ";
-  }
-  std::cout << "\b\b \n\n";
-}
-
-void DFA::printDFA(DFA &dfa) {
-  int width = 6;
-  std::cout << "\n\nThe converted DFA is:\n";
-  std::cout << "States: ";
-  for (int i : dfa.states) {
-    std::cout << i << ", ";
-  }
-
-  std::cout << "\b\b \nInput symbols: ";
-  for (char c : dfa.inputSymbols) {
-    std::cout << c << ", ";
-  }
-
-  std::cout << "\b\b \nTransition Function:\n";
-  std::cout << std::setw(width) << "";
-  for (char c : dfa.inputSymbols) {
-    std::cout << std::setw(width) << c;
-  }
-  for (int srcState : dfa.states) {
-    std::cout << std::endl << std::setw(width) << srcState;
-    for (char symbol : dfa.inputSymbols) {
-      int destState = dfa.transitionFn[srcState][symbol];
-      std::cout << std::setw(width) << destState;
-    }
-  }
-
-  std::cout << "\nStart State: " << dfa.startState;
-  std::cout << "\nFinal States: ";
-  for (int c : dfa.finalStates) {
-    std::cout << c << ", ";
-  }
-  std::cout << "\b\b \n\n";
 }
