+++ src/nfa.cpp
